refactor: ResourcePathResolver implements PropertiesFileResourcePathResolver interface
